refactor: move types to adonis-typings
