+++ .eslintignore
@@ -1,2 +1,1 @@
-src/types.ts
 lib
