+++ adonis-typings/index.ts
@@ -1,0 +1,6 @@
+/* eslint-disable @typescript-eslint/triple-slash-reference */
+
+/// <reference path="./database.ts" />
+/// <reference path="./migration.ts" />
+/// <reference path="./model.ts" />
+/// <reference path="./objectid.ts" />
