+++ adonis-typings/objectid.ts
@@ -1,0 +1,5 @@
+declare module '@ioc:Mongodb/ObjectId' {
+  import { ObjectId } from 'mongodb';
+
+  export default ObjectId;
+}
